Conclude quickstart changes with Hello World
Conclude the quickstart changes by shifting example to a simple Hello World application in order to avoid polluting the quickstart and readme with too many details. The more complex example can be made into a separate tutorial that we guide the user to if they want to learn more. 

Also included some minor verbiage changes to make the distinction between Hydra concepts and .NET concepts a bit more explicit.